Fix translate.js
Previously the file was assigning the newObj directly to student which
makes it look as though it is a new object but it is actually just
assigning it the location of student. The new code copies all the keys
and values of student into the new object allowing it to be edited
independently of the input.